Revert "Merge branch 'master' of https://github.com/Gorri4/HBV501G"
This reverts commit 8b2b9cae53c795f1863a136870b85bcf3c398c02, reversing
changes made to 4ded2e8565218cfe2c2e423b871e77c2d7ec4cc1.

+++ Hugbverk1Forrit1/src/main/java/is/hi/byrjun/model/Login.java
@@ -15,7 +15,7 @@ public class Login {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	//breta
-	private int id;
+	private float id;
 	@NotNull
 	private String userinn;
 	@NotNull
